refactor: Rename SmallAppItem to AppTile
